% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_circ.R
\name{area.circ}
\alias{area.circ}
\title{Area of a Circle}
\usage{
area.circ(radius)
}
\arguments{
\item{radius}{The only parameter needed is the radius of the circle you wish to find the area of.}
}
\value{

}
\description{
Area of a Circle
}
\examples{
area.circ(radius=5)
}
