% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{BMI Calculator: Imperial System}
\usage{
bmi(inches, pounds)
}
\arguments{
\item{inches}{Total height in inches.}

\item{pounds}{Total weigtht in pounds.}
}
\value{

}
\description{
BMI Calculator: Imperial System
}
\examples{
bmi(72,200)
}
