% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_convert.R
\name{f.convert}
\alias{f.convert}
\title{Farenheit to Celsius}
\usage{
f.convert(Farenheit)
}
\arguments{
\item{Farenheit}{Insert degrees in Farenheit}
}
\value{

}
\description{
Farenheit to Celsius
}
\examples{
f.convert(100)
}
