% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_get.R
\name{fred.get}
\alias{fred.get}
\title{Fred API}
\usage{
fred.get(api.key, series.id, obs.start)
}
\arguments{
\item{api.key}{A personal API key is necessary in accessing the Fred API.}

\item{series.id}{A series code that is required in selecting the desired data.}

\item{obs.start}{A parameter that selects the desired start time for a series.}
}
\value{

}
\description{
Fred API
}
\examples{
fred.get('YOUR API KEY','GDPC1','2020-01-01')
}
