% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{velocity}
\alias{velocity}
\title{Average Velocity}
\usage{
velocity(Distance, Time)
}
\arguments{
\item{Distance}{The total distance traveled from point A to point B.}

\item{Time}{The time it took to get from point A to point B.}
}
\value{

}
\description{
Average Velocity
}
\examples{
velocity(200,30)
}
